requirements.txt
django==4.2.17
djangorestframework==3.15.2
django-cors-headers==4.6.0
psycopg2-binary==2.9.10
python-dotenv==1.0.1
django-debug-toolbar==4.4.6
python-telegram-bot==22.0